requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qrcode[pil]==7.4.2
pillow==10.4.0
requests==2.32.3
python-multipart==0.0.9
